piematrix: use floats instead of fixed point DEG degree values for rotations
Allow passing of degrees as floats to the pie_MatRot[XYZ] functions.

This commit further helps in getting rid of fixed point math.

git-svn-id: https://warzone2100.svn.sourceforge.net/svnroot/warzone2100/trunk@11708 4a71c877-e1ca-e34f-864e-861f7616d084

+++ lib/ivis_opengl/piematrix.h
@@ -100,9 +100,9 @@ extern void pie_MatEnd(void);
 extern void pie_MATTRANS(float x, float y, float z);
 extern void pie_TRANSLATE(float x, float y, float z);
 extern void pie_MatScale(float scale);
-extern void pie_MatRotX(int x);
-extern void pie_MatRotY(int y);
-extern void pie_MatRotZ(int z);
+extern void pie_MatRotX(float x);
+extern void pie_MatRotY(float y);
+extern void pie_MatRotZ(float z);
 extern int32_t pie_RotateProject(const Vector3i *src, Vector2i *dest);
 
 //*************************************************************************
